backend/requirements.txt
absl-py==2.3.1
annotated-types==0.7.0
anyio==4.12.1
arabic-reshaper==3.0.0
asgiref==3.11.0
asn1crypto==1.5.1
astunparse==1.6.3
boto3==1.42.34
botocore==1.42.34
cachetools==6.2.5
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.3
cryptography==46.0.3
cssselect2==0.8.0
cycler==0.12.1
deprecation==2.1.0
dj-database-url==3.1.0
Django==5.2.10
django-cors-headers==4.9.0
django-ses==4.6.0
djangorestframework==3.16.1
flatbuffers==25.12.19
fonttools==4.61.1
freetype-py==2.5.1
fsspec==2026.1.0
gast==0.7.0
google-pasta==0.2.0
grpcio==1.76.0
gunicorn==24.1.1
h11==0.16.0
h2==4.3.0
h5py==3.15.1
hpack==4.1.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
jmespath==1.1.0
keras==3.13.1
kiwisolver==1.4.9
libclang==18.1.1
lxml==6.0.2
Markdown==3.10.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
ml_dtypes==0.5.4
mmh3==5.2.0
multidict==6.7.0
namex==0.1.0
numpy==2.4.1
opt_einsum==3.4.0
optree==0.18.0
oscrypto==1.3.0
packaging==26.0
pillow==12.1.0
postgrest==2.27.2
propcache==0.4.1
protobuf==6.33.4
psycopg2-binary==2.9.11
pycairo==1.29.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyHanko==0.32.0
pyhanko-certvalidator==0.29.0
pyiceberg==0.10.0
PyJWT==2.10.1
pyparsing==3.3.2
pypdf==6.6.1
pyroaring==1.0.3
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-http-client==3.3.7
PyYAML==6.0.3
realtime==2.27.2
reportlab==4.4.9
requests==2.32.5
rich==14.3.1
rlPyCairo==0.4.0
s3transfer==0.16.0
sendgrid==6.12.5
six==1.17.0
sortedcontainers==2.4.0
sqlparse==0.5.5
storage3==2.27.2
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.27.2
supabase-auth==2.27.2
supabase-functions==2.27.2
svglib==1.6.0
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.3.0
tinycss2==1.5.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
uritools==6.0.1
urllib3==2.6.3
webencodings==0.5.1
websockets==15.0.1
Werkzeug==3.1.5
whitenoise==6.11.0
wrapt==2.0.1
xhtml2pdf==0.2.17
yarl==1.22.0
